[IMP] allow context manager
The simplest possible way to implement an automatic logout with a context manager is to rely on [`contextlib.closing`][1]. The only requirement is to have a `close` function. In this case, just an alias for `logout`.

With this patch, we're able now to wrap the logged-in odoo instance with that function and get automatic logout.

See the provided test for usage example.

[1]: https://docs.python.org/3.8/library/contextlib.html#contextlib.closing